target/riscv: Make sure Andes CPUs do not enable unspecified extensions (#32)
The original code sequence would enable the default extensions of
RISC-V QEMU and allow us to dynamically configure extensions through
the command line. Although this is convenient, but it is not correct
and may potentially result in unexpected errors.

This patch fixes this issue and provides more precise configuration
for some CSRs, such as $SATP. We ensure that the satp mode of N-Series
CPUs is set to "bare".

Signed-off-by: Dylan Jhong <dylan@andestech.com>

Co-authored-by: Dylan Jhong <dylan@andestech.com>
Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/32