target/riscv: Add optional CSR tcontrol defined in Sdtrig extension target/riscv: Fix many issues since some change of v8.2 (major in cpu_env -> tcg_env) hw/intc: Replace get cpu_env function for v8.2 hw/timer: Replace get cpu_env funcrtion for v8.2
Add CSR predicate/read/write operations for tcontrol. Use new CPU
configuration "ext_sdtrig_tcontrol" to decide whether tcontrol is
implemented into CPU or not.

When any trap into M-mode is taken, tcontrol.MPTE is set to the value of
tcontrol.MTE, and tcontrol.MTE is set to 0. When mret is executed, mte
is set to the value of tcontrol.MPTE field.

Besides, when tcontrol.MTE and tcontrol.MPTE are implemented, medeleg[3]
should be hard-wired to 0.

Signed-off-by: Alvin Chang <alvinga@andestech.com>
Signed-off-by: Foxes Hung <foxes687@andestech.com>