target/riscv: Let tdata2/3 be written any value when trigger is disabled
Refer to RISC-V debug specification version 1.0-STABLE, when the trigger
is disabled, tdata2 and tdata3 of this trigger can be written with any
value that makes sense for any trigger type supported by this trigger.
The tdata2 can be written with any value. The tdata3 can be written with
supported fields, so we need to check it by textra_validate().

Signed-off-by: Alvin Chang <alvinga@andestech.com>